pyproject.toml
[tool.poetry]
name = "stable-diffusion-with-upscaler"
version = "0.1.0"
description = "End-to-end demo of using Stable Diffusion 1.4 to output an 1024x1024 image."
authors = ["SauYee Chan <sauyee.c@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.0.1"
ftfy = "^6.1.1"
braceexpand = "^0.1.7"
requests = "^2.31.0"
clip = "^0.2.0"
omegaconf = "^2.3.0"
kornia = "^0.7.0"
transformers = "4.31.0"
k-diffusion = "0.0.15"
pytorch-lightning = "1.7.7"
torchmetrics = "0.11.4"
huggingface-hub = {git = "https://github.com/huggingface/huggingface_hub"}
click = "^8.1.7"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
pre-commit = "^3.4.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
